Moved conf.py to config.py
